pyproject.toml
[project]
name = "kyykkacom"
version = "0.1.0"
description = "Website for kyykka.com"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Django>=5.1.0,<6.0.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "gunicorn>=20.0.0",
    "mysqlclient>=2.2.0",
    "Pillow",
    "PyMySQL>=1.0.0",
    "pymemcache>=4.0.0",
    "pytz>=2025.2",
    "sqlparse>=0.5.3",
    # "six>=1.16.0",
    "urllib3>=2.0.0",
    "tqdm",

]

[project.optional-dependencies]
dev = [
    "django-silk>=5.3.0",
    "django-debug-toolbar>=5.1.0",
    "Faker>=1.0.0",
    "drf-spectacular>=0.27.0",
    "pyright>=1.1.397",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["kyykka", "api", "nkl"]
